Adding default_solvers.c and fixing Bondarenko model
