+++ example_configs/benchmark_config_bondarenko.ini
@@ -1,6 +1,6 @@
 [main]
 num_threads=2
-dt_pde=0.0001
+dt_pde=0.001
 simulation_time=200
 abort_on_no_activity=false
 vm_threshold=-86.0
@@ -49,7 +49,8 @@ end_function=end_conjugate_gradient
 main_function=conjugate_gradient
 
 [ode_solver]
-dt=0.0001
+adaptive=true
+dt=0.00001
 use_gpu=true
 gpu_id=0
 library_file= shared_libs/libbondarenko_2004.so
