+++ src/models_library/ToROrd/ToRORd_fkatp_endo.h
@@ -1,7 +1,6 @@
 #ifndef MONOALG3D_MODEL_TORORD_FKATP_ENDO_H
 #define MONOALG3D_MODEL_TORORD_FKATP_ENDO_H
 
-
 #define NEQ 43
 #define INITIAL_V (-88.763800f)
