+++ src/models_library/maleckar/Maleckar2008.c
@@ -108,169 +108,7 @@ SOLVE_MODEL_ODES(solve_model_odes_cpu) {
     }
 }
 
-void solve_model_ode_cpu(real dt, real *sv, real stim_current)  {
-
-    real rY[NEQ], rDY[NEQ];
-
-    for(int i = 0; i < NEQ; i++)
-        rY[i] = sv[i];
-
-    RHS_cpu(rY, rDY, stim_current, dt);
-
-    for(int i = 0; i < NEQ; i++)
-        sv[i] = dt*rDY[i] + rY[i];
-}
-
-
-
-float __agos_factorial(int f){
-	if(f>=0 && f<2)
-		return 1.0;
-	else if(f < 0)
-		return 0.0/0.0;
-	for(int i=f-1; i>=2; i--)
-		f *= i;
-	return (float)f;
-}
-
-void solve_forward_euler_cpu_adpt(real *sv, real stim_curr, real final_time, int sv_id) {
-
-    real rDY[NEQ];
-
-    real _tolerances_[NEQ];
-    real _aux_tol = 0.0;
-    //initializes the variables
-    real dt = ode_dt[sv_id];
-    real time_new = ode_time_new[sv_id];
-    real previous_dt = ode_previous_dt[sv_id];
-
-    real edos_old_aux_[NEQ];
-    real edos_new_euler_[NEQ];
-    real *_k1__ = (real*) malloc(sizeof(real)*NEQ);
-    real *_k2__ = (real*) malloc(sizeof(real)*NEQ);
-    real *_k_aux__;
-
-    const real _beta_safety_ = 0.8;
-
-    const real __tiny_ = pow(abstol, 2.0f);
-
-    if(time_new + dt > final_time) {
-       dt = final_time - time_new;
-    }
-
-    RHS_cpu(sv, rDY, stim_curr, dt);
-    time_new += dt;
-
-    for(int i = 0; i < NEQ; i++){
-        _k1__[i] = rDY[i];
-    }
-
-    int count = 0;
-
-    int count_limit = (final_time - time_new)/min_step;
-
-    int aux_count_limit = count_limit+2000000;
-
-    if(aux_count_limit > 0) {
-        count_limit = aux_count_limit;
-    }
-
-    while(1) {
-
-        for(int i = 0; i < NEQ; i++) {
-            //stores the old variables in a vector
-            edos_old_aux_[i] = sv[i];
-            //computes euler method
-            edos_new_euler_[i] = _k1__[i] * dt + edos_old_aux_[i];
-            //steps ahead to compute the rk2 method
-            sv[i] = edos_new_euler_[i];
-        }
-
-        time_new += dt;
-        RHS_cpu(sv, rDY, stim_curr, dt);
-        time_new -= dt;//step back
-
-        double greatestError = 0.0, auxError = 0.0;
-        for(int i = 0; i < NEQ; i++) {
-            // stores the new evaluation
-            _k2__[i] = rDY[i];
-            _aux_tol = fabs(edos_new_euler_[i]) * reltol;
-            _tolerances_[i] = (abstol > _aux_tol) ? abstol : _aux_tol;
-
-            // finds the greatest error between  the steps
-            auxError = fabs(((dt / 2.0) * (_k1__[i] - _k2__[i])) / _tolerances_[i]);
-
-            greatestError = (auxError > greatestError) ? auxError : greatestError;
-        }
-        ///adapt the time step
-        greatestError += __tiny_;
-        previous_dt = dt;
-        ///adapt the time step
-        dt = _beta_safety_ * dt * sqrt(1.0f/greatestError);
-
-        if (time_new + dt > final_time) {
-            dt = final_time - time_new;
-        }
-
-        //it doesn't accept the solution
-        if ( count < count_limit  && (greatestError >= 1.0f)) {
-            //restore the old values to do it again
-            for(int i = 0;  i < NEQ; i++) {
-                sv[i] = edos_old_aux_[i];
-            }
-            count++;
-            //throw the results away and compute again
-        } else{//it accepts the solutions
-            count = 0;
-
-            if (dt < min_step) {
-                dt = min_step;
-            }
-
-            else if (dt > max_step && max_step != 0) {
-                dt = max_step;
-            }
-
-            if (time_new + dt > final_time) {
-                dt = final_time - time_new;
-            }
-
-            _k_aux__ = _k2__;
-            _k2__	= _k1__;
-            _k1__	= _k_aux__;
-
-            //it steps the method ahead, with euler solution
-            for(int i = 0; i < NEQ; i++){
-                sv[i] = edos_new_euler_[i];
-            }
-
-            if(time_new + previous_dt >= final_time) {
-                if((fabs(final_time - time_new) < 1.0e-5)) {
-                    break;
-                } else if(time_new < final_time) {
-                    dt = previous_dt = final_time - time_new;
-                    time_new += previous_dt;
-                    break;
-                } else {
-                    dt = previous_dt = min_step;
-                    time_new += (final_time - time_new);
-                    printf("Error: %lf\n", final_time - time_new);
-                    break;
-                }
-            } else {
-                time_new += previous_dt;
-            }
-
-        }
-    }
-
-	ode_dt[sv_id] = dt;
-	ode_time_new[sv_id] = time_new;
-	ode_previous_dt[sv_id] = previous_dt;
-	
-    free(_k1__);
-    free(_k2__);
-}
+#include "../default_solvers.c"
 
 void RHS_cpu(const real *sv, real *rDY, real stim_current, real dt) {
